Major refactoring
The discontinuation of the prop and prop_r types, since
those did not work at the quality and flexibility to justify
their usage. Now read-write and read-only properties
have dedicated member functions instead. This also led
to induce const-correctness into the code, which has
broken some parts of the code. Specifically, the audio
component has to be reimagined to be more consistent
with the other component structures.