Rename variables, refactor shaders and extend GUI
Started renaming some variables to be shorter and at the same time
refactored the shader code a bit. Extended the GUI so it has more
flexibility for the demo. Also discoverd that the bug introduced
in issue #4 wasn't a bug, I have overlooked that the matrix is
transposed when sent to the GPU, so the behavior is correct and
expected.

Issues resolved: #4