Improve deferred shading
Started to optimize the code for deferred shading so that instead
of performing light calculations for all lights for the entire
screen (screen width * screen height * number of lights), it
renders a quad in front of each light volume and performs the
fragment shader on only the covered pixels (quad width *
quad height * number of lights). Also added a second color
attachment to the post-processing FBO, to enable ping-ponging
between the textures for chained post-processing effects.